requirements.txt
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-CORS==3.0.10
Flask-JWT-Extended==4.5.2
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.0.3
PyMySQL==1.0.3
python-dotenv==1.0.0